✅ test(province): add test case
